_datascience/py-cardio-docx/requirements.txt
appnope==0.1.0
astroid==2.4.0
autopep8==1.5.2
backcall==0.1.0
decorator==4.4.2
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
numpy==1.18.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.5.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
pyzmq==19.0.0
six==1.14.0
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.9
wrapt==1.12.1
